Rename github.com/pkg/errors integration to package avoid name conflicts
