+++ _integrations/nrerrors/stacktrace.go
@@ -1,4 +1,4 @@
-package errors
+package nrerrors
 
 import (
 	"fmt"
